Move tests to root dir
